// @types/slides.d.ts
/**********************************************************
 * Slide data.
 *
 * @author Yuto Watanabe <yuto.w51942@gmail.com>
 * @version 1.0.0
 *
 * Copyright (C) 2021 hello-slide
 **********************************************************/

export default interface Slide {
  title: string;
  id: string;
  createDate: Date;
  lastChange: Date;
}